backend/requirements.txt
quart
quart-schema[pydantic]
quart-cors
uvicorn
pydantic>=2.0
langgraph-sdk
partial-json-parser
supabase
python-dotenv
uvloop
httptools
